chore: apply clippy --fix across the workspace (part of #84)
Machine-applicable clippy suggestions: to_vec() over iter().cloned()
.collect(), drop redundant casts/closures, is_some()/negation idioms,
move items before the test module, derive Default for AuthScheme, etc.
Tidied the AuthScheme derive (single attribute, no double blank line).
No behaviour change; all tests green.

Part of #84.

+++ crates/ruscker-docker/src/lib.rs
@@ -419,7 +419,7 @@ impl ContainerBackend for LocalDockerBackend {
                 ports
                     .iter()
                     .find(|p| p.private_port == inner_port)
-                    .and_then(|p| p.public_port.map(|n| n as u16))
+                    .and_then(|p| p.public_port.map(|n| n))
             });
             let upstream: SocketAddr = match host_port {
                 Some(p) => format!("127.0.0.1:{p}").parse().unwrap(),
